refactor: add method enum for request encapsulation fields
