+++ tomcat/src/main/java/org/apache/coyote/http11/request/Method.java
@@ -1,5 +1,7 @@
 package org.apache.coyote.http11.request;
 
+import java.util.Arrays;
+
 public enum Method {
     GET("GET"),
     POST("POST"),
@@ -9,6 +11,13 @@ public enum Method {
 
     Method(String value) {
         this.value = value;
+    }
+
+    public static Method from(String value) {
+        return Arrays.stream(Method.values())
+                .filter(method -> method.value.equals(value))
+                .findAny()
+                .orElse(null);
     }
 
     public String getValue() {
